requirements.txt
fastapi
uvicorn
litellm
qdrant-client
python-dotenv
PyPDF2
python-docx